Limit summary to 150 chars
Review Board has a limit of 300 chars. If a summary is bigger
it causes an error because the summary seems modified even it
is just cut off.

Also most times there is no need for so big summary. It's just
because users don't understand summary and description.

ERROR: You attempted to publish a review request without any modifications. (HTTP 400, API Error 211)