Iterate all_items for get_file_attachments
Otherwise we will support 25 images only because
you cannot remove image 26 or higher from that
Review Request.

Also we should not try to delete an image that we
do not know anymore. This could only happen if
someone terminates an active push and the Review
Request has an incomplete draft.