Do not fetch --debug for head check
If an existing head used transplant the extra_data
has a binary field "transplant_source" and python
throws an encoding exception.
As we cannot change existing changesets in the repository
we should avoid that as we do not need it here.

New changesets needs to be adjusted and should use graft.